Restructure Project
Rename src to lib and use es5 as target of the es6 to es5 compilation.
Also transpile unit tests to allow debugging in webstorm.
