+++ .gitignore
@@ -2,4 +2,4 @@
 node_modules
 graph.png
 coverage
-dist
+es5
