+++ lib/cfg/cfg-builder.js
@@ -47,13 +47,6 @@ const cfgVisitor = {
 		this.cfg.connectIfNotFound(forGraphNode, BRANCHES.FALSE, successor(path));
 		this.cfg.connectIfNotFound(forGraphNode, BRANCHES.TRUE, path.node.body);
 		this.cfg.connectIfNotFound(path.node.update, BRANCHES.UNCONDITIONAL, forGraphNode);
-
-
-		// this.cfg.connectIfNotFound(path.node, BRANCHES.TRUE, path.node.body);
-		// this.cfg.connectIfNotFound(path.node, BRANCHES.FALSE, successor(path));
-
-		// for (x in array) { body }
-		// for (x of array) { body }
 	}
 };
