+++ lib/cfg/edge.js
