+++ lib/cfg/node.js
