+++ lib/cfg/plot.js
