+++ lib/cfg/successor.js
@@ -29,7 +29,7 @@ function getParentLoopFor (path) {
  * @param path {NodePath} the path for which the successor should be determined
  * @returns {NodePath} the successor node or null if the direct successor of this node is the EOF (end of file, or end of program).
  */
-export default function successor (path) {
+export function successor (path) {
 	const parent = path.parentPath;
 
 	if (parent) {
@@ -44,18 +44,23 @@ export default function successor (path) {
 		}
 	}
 
+	// the successor of the break statement is the successor of the containing loop
 	if (path.isBreakStatement()) {
 		return successor(getParentLoopFor(path));
 	}
 
+	// the successor of a continue statement is the containing loop
 	if (path.isContinueStatement()) {
 		return getParentLoopFor(path);
 	}
 
+	// if the path is a program, then we have reached the end -> EOF
 	if (path.isProgram()) {
 		return null;
 	}
 
+	// for all other statements, the successor is the following sibling node.
+	// If the node has no following sibling node, then we reached the end and need to continue on the upper level
 	let sibling = path.getSibling(path.key + 1);
 	if (!sibling.node) {
 		return successor(path.parentPath);
@@ -63,3 +68,5 @@ export default function successor (path) {
 
 	return fallThrough(sibling);
 }
+
+export default successor;
