+++ lib/cli.js
@@ -6,7 +6,7 @@ import Configuration from "./configuration";
 import Context from "./context";
 import {inferProgram} from "./infer";
 
-cliProgram.version(require("../package.json").version)
+cliProgram.version("0.0.1")
     .option("-f, --file <path>", "the entry file for the application to validate.")
     .parse(process.argv);
