+++ lib/configuration.js
