+++ lib/infer.js
