+++ lib/module-resolution/node-module-resolution.js
