+++ lib/semantic-model/program.js
