+++ lib/semantic-model/scope.js
