+++ lib/semantic-model/source-file.js
