+++ lib/semantic-model/symbol.js
