+++ lib/validator.js
@@ -3,8 +3,6 @@
 import fs from "fs";
 import resolve from "resolve";
 import * as babylon from "babylon";
-import traverse from "babel-traverse";
-import * as t from "babel-types";
 
 /**
  * Validates the application in the given entry file with all it's dependencies.
