+++ test/program.specs.js
@@ -1,3 +1,3 @@
-import "../src/infer";
+import "../lib/infer";
 
 // empty test that is currently only used to fix the test coverage
